requirements.txt
pandas==2.2.2
numpy==1.26.4
pyspark==3.5.0
kafka-python==2.0.0
websocket-client==1.8.0
notebook
apache-airflow==2.10.0
requests
psycopg2-binary
streamlit